backend/requirements.txt
# FastAPI and dependencies
fastapi==0.109.1
uvicorn[standard]==0.27.0
pydantic==2.6.0
pydantic-settings==2.1.0

# Plex integration
PlexAPI==4.15.6

# Database
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9  # PostgreSQL driver - ADDED FOR TIMESCALEDB

# HTTP client
httpx==0.26.0
aiohttp==3.9.1

# Utilities
python-dotenv==1.0.0
python-multipart==0.0.6

# Logging
loguru==0.7.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
